requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.3

# Cache / Queue
redis==5.1.1
celery==5.4.0

# Slack
slack-bolt==1.21.0
slack-sdk==3.33.0

# HTTP Client
httpx==0.27.2

# AI
anthropic==0.40.0

# Auth / Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
itsdangerous==2.2.0

# Config
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Utilities
tenacity==9.0.0
structlog==24.4.0
rich==13.9.4

# Dev
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-httpx==0.32.0
httpx==0.27.2
